install.wrapper.c: flesh out argument-taking args
We can't rely on getopt_long returning '?' to skip arguments
that take a subsequent argument in short form.  This can result
in a miscalculation of "first".  For example,

  xattr-install -m 0555 foo bar

when bar is a directory will fail due to there being no file "0555"
without this change.

It's not clear to me what is exactly right for the optional argument
'Z'.  In coreutils the optstring for it is just 'Z' but I put 'Z::'
for it since the manpage says that's how we do for optional arguments.
But maybe this is wrong.  It's a stupid semi-deprecated corner-case,
but.... I'm sure somebody will figure a way to get bit by it if
I'm wrong, sooner or later :(

I hate optarg.  Fuck you, Mr. FancyAPI!  Long live switch (*c++)!

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>